Try removing the N parameter from Signal
Keeping track of the number of channels makes type stability very tough.  It make some methods easier, but the only thing I've found it useful for dispatch is display.

Note that in other domains, e.g., Audio, this may be more important.  We can always add it back in if it will be useful.