Major refactor: Signals are multidimensional arrays
This generalizes Signals to be much simpler and more powerful. Instead of being Vectors of Vectors, they are now complete multidimensional arrays that return vectors. Their storage is an N+1 array.  This allows for much less reliance on type inference, but requires that all elements be of the same type.

Still todo: RaggedSignals with a variable number of timepoints.